network-grpc: Move the service implementation into a private module
The NodeService implementation type and its infrastructure do not need
to be public, and that code is not coupled to Server and Connection.
It makes sense to move it to a private module and leave mod server
to the server API.
